refactor: split DojoViewModelTest to remove LargeClass suppression
DojoViewModelTest had grown to 1042 lines behind an inline
@Suppress("LargeClass"), the only suppression backstopping it since the
repo has no detekt baseline and maxIssues is 0. Split its 43 tests into
four behavior-focused classes (core flow, quota math, due-count/empty-state
reactivity, undo) so detekt's LargeClass rule passes without suppression.
